requirements.txt
requests==2.31.0
pandas==2.1.4
pymysql==1.1.0
sqlalchemy==2.0.23
python-dotenv==1.0.0